moved management webapp classes to a new support module
