+++ cas-management-webapp-support/src/main/java/org/jasig/cas/services/web/beans/RegisteredServiceAttributeReleasePolicyViewBean.java
