+++ cas-management-webapp-support/src/test/java/org/jasig/cas/WiringTests.java
@@ -23,9 +23,9 @@ public class WiringTests {
     public void setUp() {
         applicationContext = new XmlWebApplicationContext();
         applicationContext.setConfigLocations(
-                "file:src/main/webapp/WEB-INF/cas-management-servlet.xml",
-                "file:src/main/webapp/WEB-INF/managementConfigContext.xml",
-        "file:src/main/webapp/WEB-INF/spring-configuration/*.xml");
+                "classpath:/cas-management-servlet.xml",
+                "classpath:/managementConfigContext.xml",
+        "classpath:/spring-configuration/*.xml");
         applicationContext.setServletContext(new MockServletContext(new ResourceLoader() {
             @Override
             public Resource getResource(final String location) {
