+++ cas-management-webapp-support/src/test/resources/log4j2.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
-<Configuration>
+<!-- Specify the refresh internal in seconds. -->
+<Configuration monitorInterval="60">
     <Appenders>
         <Console name="console" target="SYSTEM_OUT">
             <PatternLayout pattern="%d %p [%c] - &lt;%m&gt;%n"/>
@@ -9,15 +10,20 @@
             <PatternLayout pattern="%d %p [%c] - %m%n"/>
             <Policies>
                 <OnStartupTriggeringPolicy />
-                <SizeBasedTriggeringPolicy size="1 MB"/>
+                <SizeBasedTriggeringPolicy size="512 KB"/>
                 <TimeBasedTriggeringPolicy />
             </Policies>
         </RollingFile>
     </Appenders>
     <Loggers>
         <Logger name="org.springframework" level="warn" />
-        <Logger name="org.jasig" level="info">
+        <Logger name="org.springframework.webflow" level="warn" />
+        <Logger name="org.springframework.web" level="warn" />
+        <Logger name="org.springframework.security" level="warn" />
+
+        <Logger name="org.jasig" level="debug" additivity="false">
             <AppenderRef ref="cas-management"/>
+            <AppenderRef ref="console"/>
         </Logger>
         <Logger name="org.jasig.inspektr.audit.support.Slf4jLoggingAuditTrailManager" level="info">
             <AppenderRef ref="cas-management"/>
