+++ cas-management-webapp-support/src/test/resources/spring-configuration/applicationContext.xml
