+++ cas-management-webapp-support/src/test/resources/spring-configuration/propertyFileConfigurer.xml
@@ -14,5 +14,5 @@
         can be moved between tiers without modification.
     </description>
 
-    <context:property-placeholder location="/WEB-INF/cas-management.properties"/>
+    <context:property-placeholder location="classpath:/cas-management.properties"/>
 </beans>
