+++ cas-management-webapp/build.gradle
@@ -13,98 +13,8 @@ war {
     }
 }
 
-
-
-
 dependencies {
-    compile(group: 'org.springframework.security', name: 'spring-security-core', version: springSecVersion) {
-        exclude(module: 'spring-core')
-        exclude(module: 'spring-beans')
-        exclude(module: 'spring-context')
-        exclude(module: 'spring-aop')
-        exclude(module: 'spring-expression')
-        exclude(module: 'spring-web')
-    }
-    compile(group: 'org.jasig.inspektr', name: 'inspektr-audit', version: inspektrVersion) {
-        exclude(module: 'slf4j-api')
-        exclude(module: 'validation-api')
-        exclude(module: 'aspectjrt')
-    }
-    compile(group: 'org.springframework.security', name: 'spring-security-web', version: springSecVersion) {
-        exclude(module: 'spring-core')
-        exclude(module: 'spring-beans')
-        exclude(module: 'spring-context')
-        exclude(module: 'spring-aop')
-        exclude(module: 'spring-expression')
-        exclude(module: 'spring-web')
-    }
-    compile group: 'org.springframework', name: 'spring-aop', version: springVersion
-    compile(group: 'org.springframework.webflow', name: 'spring-binding', version: springWebflowVersion) {
-        exclude(module: 'commons-logging')
-        exclude(module: 'spring-beans')
-        exclude(module: 'spring-context')
-        exclude(module: 'spring-core')
-        exclude(module: 'spring-expression')
-        exclude(module: 'spring-web')
-    }
-    compile project(':cas-server-core-services')
-    compile project(':cas-server-core-authentication')
-    compile project(':cas-server-core')
-    compile project(':cas-server-core-logout')
-    compile project(':cas-server-core-logging')
-    compile project(':cas-server-core-tickets')
-    compile project(':cas-server-core-audit')
-    compile project(':cas-server-support-oauth')
-    compile(group: 'org.jasig.cas.client', name: 'cas-client-core', version: casClientVersion) {
-        exclude(module: 'servlet-api')
-        exclude(module: 'log4j-over-slf4j')
-        exclude(module: 'bcprov-jdk15')
-        exclude(module: 'slf4j-api')
-    }
-    compile group: 'org.springframework', name: 'spring-context-support', version: springVersion
-    compile group: 'org.springframework', name: 'spring-expression', version: springVersion
-    compile(group: 'org.springframework', name: 'spring-core', version: springVersion) {
-        exclude(module: 'commons-logging')
-    }
-    compile group: 'org.springframework', name: 'spring-beans', version: springVersion
-    compile group: 'org.springframework', name: 'spring-webmvc', version: springVersion
-    compile group: 'javax.servlet', name: 'jstl', version: javaxJstlVersion
-    runtime(group: 'org.jasig.inspektr', name: 'inspektr-support-spring', version: inspektrVersion) {
-        exclude(module: 'slf4j-api')
-        exclude(module: 'validation-api')
-        exclude(module: 'spring-web')
-        exclude(module: 'spring-webmvc')
-        exclude(module: 'spring-tx')
-        exclude(module: 'spring-beans')
-        exclude(module: 'spring-jdbc')
-        exclude(module: 'aspectjrt')
-        exclude(module: 'validation-api')
-    }
-    runtime(group: 'org.springframework.security', name: 'spring-security-cas', version: springSecVersion) {
-        exclude(module: 'commons-logging')
-        exclude(module: 'spring-core')
-        exclude(module: 'spring-beans')
-        exclude(module: 'spring-context')
-        exclude(module: 'spring-aop')
-        exclude(module: 'spring-expression')
-        exclude(module: 'spring-web')
-        exclude(module: 'cas-client-core')
-    }
-    runtime(group: 'org.springframework.security', name: 'spring-security-config', version: springSecVersion) {
-        exclude(module: 'commons-logging')
-        exclude(module: 'spring-core')
-        exclude(module: 'spring-beans')
-        exclude(module: 'spring-context')
-        exclude(module: 'spring-aop')
-        exclude(module: 'spring-expression')
-        exclude(module: 'spring-web')
-    }
-    runtime group: 'org.apache.logging.log4j', name: 'log4j-web', version: log4jVersion
-    runtime group: 'taglibs', name: 'standard', version: taglibsVersion
-    runtime(group: 'org.hibernate', name: 'hibernate-validator', version: hibernateValidatorVersion) {
-        exclude(module: 'slf4j-api')
-        exclude(module: 'jboss-logging')
-    }
+    runtime project(':cas-management-webapp-support')
 }
 
 configurations {
