--- cas-management-webapp/src/main/java/org/jasig/cas/services/web/package.html
@@ -1,5 +1,0 @@
-<html>
-<body>
-<p>This package contains all the controllers and validator for the services management.</p>
-</body>
-</html>
