+++ settings.gradle
@@ -53,3 +53,4 @@ include ':cas-server-webapp-actions'
 include ':cas-server-webapp-support'
 include ':cas-server-webapp'
 include ':cas-management-webapp'
+include ':cas-management-webapp-support'
